refactor: 3b (commands) — extract cmd_init + cmd_validate to their own modules
Move cmd_init (+ starter helpers _prompt_starter_type, _starter_dir_slug,
_resolve_starter_dir) to cli/cmd_init.py and cmd_validate to cli/cmd_validate.py.
govkit re-exports cmd_init/cmd_validate for main() dispatch, plus
_resolve_starter_dir and read_govkit_level as self-aliases (now test-only, their
last internal caller cmd_init moved out). govkit.py: 840 -> 740 lines.
No behavior change; 807 tests green.
